requirements.txt
pocketbase
requests
python-dotenv